"new" install - loading settings
If the user manually enters a path in the twinBASIC location textbox and then moves away from that textbox, the folder will be validated. If it isn't found, they will be asked if it should be created. This will allow those that do not have a tB folder already to use this utility to download and unzip it.

Settings are no longer wiped if one setting is a problem.